Change Price from money to int

+++ PhoneSaleAPI/Models/BillDetail.cs
@@ -8,8 +8,8 @@ namespace PhoneSaleAPI.Models
         public string BillId { get; set; } = null!;
         public string ProductId { get; set; } = null!;
         public int? Amount { get; set; }
-        public decimal? Price { get; set; }
-        public decimal? Total { get; set; }
+        public int? Price { get; set; }
+        public int? Total { get; set; }
 
         public virtual Bill Bill { get; set; } = null!;
         public virtual Product Product { get; set; } = null!;
